Fix engineering nits per review
1. Fix frame counter types: both cd /workspace && git add -A && git commit -m "Fix engineering nits per review

1. Fix frame counter types: both $1842 and $09DA are uint16 (not uint32)
2. Fix --load-state argv parsing: accept before/after optional 'predict'
   - sm_rev_predict predict --load-state FILE
   - sm_rev_predict --load-state FILE predict
   - sm_rev_predict --load-state FILE
3. Add Python CLI test to 'make mini-test' (no ROM)
4. Clarify C test: frame-0 hydrate only, NOT grounded-walk residual
   - Grounded-walk R(τ) vs SuperMetroidEnv is later work
5. Omit is_dead/is_game_over from JSON (hardcoded false, not measured)
   - Document in code comments only

All tests pass. CLI arg parsing works for SmRevClient [bin, 'predict'] and
all --load-state positions. Mini-test now includes Python CLI test."842 and --9DA are uint16 (not uint32)
2. Fix --load-state argv parsing: accept before/after optional 'predict'
   - sm_rev_predict predict --load-state FILE
   - sm_rev_predict --load-state FILE predict
   - sm_rev_predict --load-state FILE
3. Add Python CLI test to 'make mini-test' (no ROM)
4. Clarify C test: frame-0 hydrate only, NOT grounded-walk residual
   - Grounded-walk R(τ) vs SuperMetroidEnv is later work
5. Omit is_dead/is_game_over from JSON (hardcoded false, not measured)
   - Document in code comments only

All tests pass. CLI arg parsing works for SmRevClient [bin, 'predict'] and
all --load-state positions. Mini-test now includes Python CLI test.

+++ src/mini/mini_predict.c
@@ -44,13 +44,13 @@ MiniTrajectoryFrame MiniCaptureTrajectoryFrame(const MiniGameState *state, int f
   // Energy: read from g_ram $09C2
   result.energy = samus_health;                // g_ram $09C2
   
-  // Death and game over state (mini doesn't track death state yet, future work)
+  // Death and game over state: not yet implemented in Mini
+  // These fields are NOT emitted in JSON output (see predict_cli.c)
   result.is_dead = false;
   result.is_game_over = false;
   
-  // Frame counters: read from g_ram for observation tuple
-  // Avoid naming conflicts with variables.h macros by using generic names
-  result.frame_counter_1 = *(uint32*)(g_ram + kWramAddr_FrameCounter1);
+  // Frame counters: read from g_ram (both uint16 per locked M-E map)
+  result.frame_counter_1 = *(uint16*)(g_ram + kWramAddr_FrameCounter1);
   result.frame_counter_2 = *(uint16*)(g_ram + kWramAddr_FrameCounter2);
   
   // Enemy tracking: capture active enemies from MiniSim (pixel x/y only)
